insurancecopay as float32
Reviewer       :
Code review id :
Tests run      :
Issue fixed    :

To fix error:
2022/01/13 15:08:10 Error unmarshaling response body: json: cannot unmarshal number 80.34 into Go struct field .appointments.appointmentcopay.insurancecopay of type int

+++ athenahealth/appointments.go
@@ -79,7 +79,7 @@ type BookedAppointment struct {
 	AppointmentCopay struct {
 		CollectedForOther       int `json:"collectedforother"`
 		CollectedForAppointment int `json:"collectedforappointment"`
-		InsuranceCopay          int `json:"insurancecopay"`
+		InsuranceCopay          float32 `json:"insurancecopay"`
 	} `json:"appointmentcopay"`
 	AppointmentNotes []struct {
 		DisplayOnSchedule bool   `json:"displayonschedule"`
